+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 2024/10/01
+
+* Modified so that the gizmo can be operated even if the target Node3D and the collision of the transform gizmo overlap.
+
 ## 2024/09/20
 
 * Transform gizmos are now drawn at the top level.